chore(licenses): move the audit into scripts/lib so the test can import it
The unit tests imported scripts/check-licenses.mjs directly, and that file
carries a `#!/usr/bin/env node` shebang. vitest cannot parse a module with a
shebang, so the whole suite failed to load with 'SyntaxError: Invalid or
unexpected token' and the gate never ran.

Split along the line the rest of scripts/ already uses: scripts/lib/*.mjs is
importable and has no shebang, scripts/*.mjs is the CLI entry point. The audit
moves to scripts/lib/license-audit.mjs; check-licenses.mjs keeps the shebang
and does argument parsing and reporting only.

No behaviour change — same closure, same policy, same exit codes.
